+++ common/src/main/java/net/spell_engine/internals/SpellCastAction.java
@@ -1,5 +1,0 @@
-package net.spell_engine.internals;
-
-public enum SpellCastAction {
-    START, CHANNEL, RELEASE
-}